Fix sideways pipe hitbox collision (#730)
* Fixed sideways pipe hitbox collision

This fixes a bug where players would clip the top edge of a sideways pipe's hitbox and get counted as being on the floor despite being above the pipe, thus allowing them to get warped into the pipe from there.

* Fix to prevent entering pipes from too low

this implements a simple distance check from the player to the hitbox to ensure the player can only ever enter the pipe when they're close enough in proximity to the pipe, preventing them from entering from a tile too low.

* co mnment

:earnest: