Various ROM verification screen improvements (#682)
* Added file prompt button for selecting rom and additional error checks

Updated the rom verification screen with clarified errors and a slight bug fix to the error text.

Along with this, there is now a button on the rom verification screen to select a file directly, defaulting to an array of NES rom extensions, and a proper theme set for buttons, which shouldn't interfere with other themes.

* Updated label and repositioned elements

Label now properly mentions either selecting the rom file using the button or dragging it onto the window. Button also clarified a little more clearly

* final touches

* bro

* i hate labels